third_party/WebKit/LayoutTests/resize-observer/observe-expected.txt

CONSOLE ERROR: ResizeObserver loop limit exceeded
This is a testharness.js-based test.
PASS guard 
PASS test0: simple observation 
PASS test1: multiple observation on same element trigger only one 
FAIL test2: throw exception when observing non-element Test bug: need to pass exception to assert_throws()
PASS test3: disconnect stops all notifications 
PASS test4: unobserve target stops notifications, unobserve non-observed does nothing 
PASS test5: observe img 
PASS test6: iframe notifications 
PASS test7: callback.this 
Harness: the test ran to completion.
